Move to team module (#202)
* move team related functions to team module

* ilnt
