+++ challengeutils/utils.py
@@ -286,85 +286,6 @@ def change_all_submission_status(syn, evaluationid, submission_status='SCORED',
         syn.store(status)
 
 
-class NewUserProfile(synapseclient.team.UserProfile):
-    '''
-    Create new user profile that makes Userprofiles hashable
-    SYNPY-879
-    '''
-    def __hash__(self):
-        return int(self['ownerId'])
-
-
-def _get_team_set(syn, team):
-    '''
-    Helper function to return a set of usernames
-
-    Args:
-        syn: Synapse object
-        team: Synapse team id, name or object
-
-    Returns:
-        Set of synapse user profiles in team
-    '''
-    members = syn.getTeamMembers(team)
-    members_set = set(NewUserProfile(**member['member']) for member in members)
-    return members_set
-
-
-def team_members_diff(syn, a, b):
-    '''
-    Calculates the diff between teama and teamb
-
-    Args:
-        syn: Synapse object
-        a: Synapse Team id or name
-        b: Synapse Team id or name
-
-    Returns:
-        Set of synapse user profiles in teama but not in teamb
-    '''
-    uniq_teama_members = _get_team_set(syn, a)
-    uniq_teamb_members = _get_team_set(syn, b)
-    members_not_in_teamb = uniq_teama_members.difference(uniq_teamb_members)
-    return members_not_in_teamb
-
-
-def team_members_intersection(syn, a, b):
-    '''
-    Calculates the intersection between teama and teamb
-
-    Args:
-        syn: Synapse object
-        a: Synapse Team id or name
-        b: Synapse Team id or name
-
-    Returns:
-        Set of synapse user profiles that belong in both teams
-    '''
-    uniq_teama_members = _get_team_set(syn, a)
-    uniq_teamb_members = _get_team_set(syn, b)
-    intersect_members = uniq_teama_members.intersection(uniq_teamb_members)
-    return intersect_members
-
-
-def team_members_union(syn, a, b):
-    '''
-    Calculates the union between teama and teamb
-
-    Args:
-        syn: Synapse object
-        a: Synapse Team id or name
-        b: Synapse Team id or name
-
-    Returns:
-        Set of a combination of synapse user profiles from both teams
-    '''
-    uniq_teama_members = _get_team_set(syn, a)
-    uniq_teamb_members = _get_team_set(syn, b)
-    union_members = uniq_teama_members.union(uniq_teamb_members)
-    return union_members
-
-
 def _check_date_range(date_str, start_datetime, end_datetime):
     '''
     Helper function to check if the date is within range
@@ -518,16 +439,3 @@ def delete_submission(syn, submissionid):
     """
     sub = syn.getSubmission(submissionid, downloadFile=False)
     syn.delete(sub)
-
-
-def remove_team_member(syn, team, user):
-    """Removes team member
-
-    Args:
-        syn: Synapse object
-        team: synaspeclient.Team or its id
-        user: synapseclient.UserProfile or its id
-    """
-    teamid = id_of(team)
-    userid = id_of(user)
-    syn.restDELETE(f"/team/{teamid}/member/{userid}")
